fapi2: migrate to use targeting over pdbg
This patch updates methods in hwpf target and plat_error to use the
Targeting library instead of pdbg. With this change, device tree access
is now handled through the Targeting APIs, replacing pdbg-based logic.

This is patchset 1 in the migration series. Future commits will extend
this approach to remaining methods and components.

Signed-off-by: Marri Devender Rao <devenrao@in.ibm.com>
Change-Id: Ibda2e37ec0219a284d293c277579e60ee9d5c8ee